refactor(web): extract backtesting request builders
